cleanup imports from scraperAPI.database
